# Node/swagger/spec.json
{"openapi":"3.0.0","info":{"title":"BellPou API","version":"1.0.0"},"paths":{"/client":{"get":{"tags":["Client"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The number of items to skip before starting to collect the result set"},{"in":"query","name":"size","schema":{"type":"integer"},"description":"The numbers of items to return"}],"responses":{"200":{"$ref":"#/components/responses/ClientFound"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}},"post":{"tags":["Client"],"requestBody":{"$ref":"#/components/requestBodies/ClientToAdd"},"responses":{"201":{"$ref":"#/components/responses/ClientAdded"},"400":{"$ref":"#/components/responses/InputError"},"500":{"description":"Server error"}}},"patch":{"tags":["Client"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/ClientToUpdate"},"responses":{"204":{"$ref":"#/components/responses/ClientUpdated"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/personalReward":{"get":{"tags":["PersonalReward"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The number of items to skip before starting to collect the result set"},{"in":"query","name":"size","schema":{"type":"integer"},"description":"The numbers of items to return"}],"responses":{"200":{"$ref":"#/components/responses/PersonalRewardsFound"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}},"post":{"tags":["PersonalReward"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/PersonalRewardToAdd"},"responses":{"201":{"$ref":"#/components/responses/PersonalRewardAdded"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}},"delete":{"tags":["PersonalReward"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/PersonalRewardToDelete"},"responses":{"204":{"$ref":"#/components/responses/PersonalRewardDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Server error"}}}},"/ping":{"get":{"tags":["Ping"],"responses":{"200":{"description":"API reachable","content":{"text/plain":{"schema":{"type":"string","example":"pong"}}}}}}},"/report/scanQR":{"post":{"tags":["Report"],"requestBody":{"$ref":"#/components/requestBodies/ReportToAdd"},"responses":{"201":{"$ref":"#/components/responses/ReportAdded"},"400":{"$ref":"#/components/responses/InputError"},"500":{"description":"Server error"}}}},"/reward/{id}":{"get":{"tags":["Reward"],"parameters":[{"name":"id","description":"ID of a reward","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/RewardFound"},"404":{"description":"Reward not found"},"500":{"description":"Server Error"}}}},"/reward/":{"get":{"tags":["Reward"],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The number of items to skip before starting to collect the result set"},{"in":"query","name":"size","schema":{"type":"integer"},"description":"The numbers of items to return"}],"responses":{"200":{"$ref":"#/components/responses/RewardsFound"},"400":{"$ref":"#/components/responses/InputError"},"500":{"description":"Server error"}}}},"/reward":{"post":{"tags":["Reward"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/RewardToAdd"},"responses":{"201":{"$ref":"#/components/responses/RewardAdded"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}},"delete":{"tags":["Reward"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/RewardToDelete"},"responses":{"200":{"$ref":"#/components/responses/RewardDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}},"patch":{"tags":["Reward"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/RewardToUpdate"},"responses":{"204":{"$ref":"#/components/responses/RewardUpdated"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"Reward doesn't exist"},"500":{"description":"Server error"}}}},"/trash":{"get":{"tags":["Trash"],"responses":{"200":{"$ref":"#/components/responses/TrashsFound"},"400":{"$ref":"#/components/responses/InputError"},"500":{"description":"Server error"}},"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The number of items to skip before starting to collect the result set"},{"in":"query","name":"size","schema":{"type":"integer"},"description":"The numbers of items to return"}]},"patch":{"tags":["Trash"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/TrashToUpdate"},"responses":{"204":{"$ref":"#/components/responses/TrashUpdated"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"Trash doesn't exist"},"500":{"description":"Server error"}}},"post":{"tags":["Trash"],"requestBody":{"$ref":"#/components/requestBodies/TrashToAdd"},"responses":{"201":{"$ref":"#/components/responses/TrashAdded"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}},"delete":{"tags":["Trash"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/TrashToDelete"},"responses":{"200":{"$ref":"#/components/responses/TrashDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}}},"/trash/empty":{"post":{"tags":["Trash"],"requestBody":{"$ref":"#/components/requestBodies/TrashToEmpty"},"responses":{"201":{"$ref":"#/components/responses/TrashEmptied"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"Trash not found"},"500":{"description":"Server error"}}}},"/user/login":{"post":{"tags":["User"],"requestBody":{"$ref":"#/components/requestBodies/Login"},"responses":{"201":{"$ref":"#/components/responses/Logged"},"404":{"description":"The user is not registered"},"500":{"description":"Server error"}}}},"/user":{"get":{"tags":["User"],"security":[{"bearerAuth":[]}],"responses":{"200":{"$ref":"#/components/responses/UserFound"},"401":{"description":"No session is available"},"500":{"description":"Server error"}}}},"/vendor":{"get":{"tags":["Vendor"],"responses":{"200":{"$ref":"#/components/responses/VendorsFound"},"400":{"$ref":"#/components/responses/InputError"},"500":{"description":"Server error"}},"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"The number of items to skip before starting to collect the result set"},{"in":"query","name":"size","schema":{"type":"integer"},"description":"The numbers of items to return"}]},"post":{"tags":["Vendor"],"requestBody":{"$ref":"#/components/requestBodies/VendorToAdd"},"responses":{"201":{"$ref":"#/components/responses/VendorAdded"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}},"delete":{"tags":["Vendor"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/VendorToDelete"},"responses":{"200":{"$ref":"#/components/responses/VendorDeleted"},"400":{"$ref":"#/components/responses/ErrorJWT"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Server error"}}},"patch":{"tags":["Vendor"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/VendorToUpdate"},"responses":{"204":{"$ref":"#/components/responses/VendorUpdated"},"400":{"description":"client error","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/responses/ErrorJWT"},{"$ref":"#/components/responses/InputError"}]}}}},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"Vendor doesn't exist"},"500":{"description":"Server error"}}}}},"components":{"schemas":{"Client":{"type":"object","properties":{"id":{"type":"integer"},"first_name":{"type":"string"},"last_name":{"type":"string"},"birth_date":{"type":"string","format":"date"},"nb_throins":{"type":"integer"},"email":{"type":"string","format":"email"},"nb_bad_report":{"type":"integer","minimum":0},"is_banned":{"type":"boolean"}},"required":["first_name","last_name","birth_date","nb_throins","email","nb_bad_report","is_banned"]},"PersonalReward":{"type":"object","properties":{"id":{"type":"integer"},"discount_code":{"type":"string"},"exp_date":{"type":"string","format":"date"},"is_used":{"type":"boolean"},"client_id":{"type":"integer"},"reward_id":{"type":"integer"}},"required":["discount_code","exp_date","is_used","client_id","reward_id"]},"Report":{"type":"object","properties":{"trash_id":{"type":"integer"},"client_id":{"type":"integer"}},"required":["client_id","trash_id"]},"Reward":{"type":"object","properties":{"id":{"type":"integer"},"name_fr":{"type":"string"},"name_en":{"type":"string"},"description_fr":{"type":"string"},"description_en":{"type":"string"},"throins_cost":{"type":"integer"},"real_cost":{"type":"number","format":"float"},"vendor_id":{"type":"integer"}},"required":["name_fr","name_en","description_fr","description_en","throins_cost","real_cost","vendor_id"]},"Trash":{"type":"object","properties":{"id":{"type":"integer"},"is_full":{"type":"boolean"},"nb_alerts":{"type":"integer"},"last_empty":{"type":"string","format":"date"},"qr_code":{"type":"string"},"position_id":{"type":"integer","$ref":"#/components/schemas/Position"}},"required":["is_full","nb_alerts","last_empty","qr_code","position_id"]},"Vendor":{"type":"object","properties":{"id":{"type":"integer"},"name_fr":{"type":"string"},"name_en":{"type":"string"},"description_fr":{"type":"string"},"description_en":{"type":"string"},"throins_cost":{"type":"integer"},"real_cost":{"type":"number","format":"float"},"position_id":{"type":"integer","$ref":"#/components/schemas/Position"}},"required":["name_fr","name_en","description_fr","description_en","throins_cost","real_cost","position_id"]},"Manager":{"type":"object","properties":{"id":{"type":"integer"},"first_name":{"type":"string"},"last_name":{"type":"string"},"birth_date":{"type":"string","format":"date"},"nb_throins":{"type":"integer"},"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}},"required":["first_name","last_name","birth_date","nb_throins","email","password"]},"Position":{"type":"object","properties":{"id":{"type":"integer"},"coordinate_x":{"type":"integer","format":"float"},"coordinate_y":{"type":"string","format":"float"}},"required":["coordinate_x","coordinate_y"]}},"responses":{"ClientAdded":{"description":"The client has been added"},"ClientUpdated":{"description":"The Client has been updtated"},"ClientDeleted":{"description":"The client has been deleted"},"PersonalRewardsFound":{"description":"return all the personal reward of the user session","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonalReward"}}}}},"PersonalRewardDeleted":{"description":"The personalReward has been deleted"},"ReportAdded":{"description":"The report has been added"},"RewardFound":{"description":"return a Reward","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reward"}}}},"RewardsFound":{"description":"return un array of Rewards","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reward"}}}}},"RewardDeleted":{"description":"The Reward has been deleted"},"RewardUpdated":{"description":"The reward has been updtated"},"TrashFound":{"description":"return a Trash","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trash"}}}},"TrashAdded":{"description":"The trash has been added"},"TrashUpdated":{"description":"The trash has been updtated"},"TrashDeleted":{"description":"The Trash has been deleted"},"TrashEmptied":{"description":"The trash has been emptied"},"Logged":{"description":"The used is logged !"},"UserFound":{"description":"Return the user in session","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}}},"VendorsFound":{"description":"All the vendors has been received"},"VendorFound":{"description":"return a Vendor","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vendor"}}}},"VendorAdded":{"description":"The vendor has been added"},"VendorDeleted":{"description":"The Vendor has been deleted"},"VendorUpdated":{"description":"The vendor has been updtated"},"mustBeManager":{"description":"The requested action can only be performed by a manager"},"ErrorJWT":{"description":"The JWT is not valid"},"MissingJWT":{"description":"No JWT"},"InputError":{"description":"Input error"}},"requestBodies":{"ClientToAdd":{"content":{"application/json":{"schema":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"birth_date":{"type":"string","format":"date"},"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}},"required":["first_name","last_name","birth_date","email","nb_bad_report","is_banned"]}}}},"ClientToUpdate":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Client"}}}},"ClientToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"}},"required":["id"]}}}},"PersonalRewardToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"}},"required":["id"]}}}},"ReportToAdd":{"content":{"application/json":{"schema":{"type":"object","properties":{"qr_code":{"type":"string"}},"required":["qr_code"]}}}},"RewardToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"}},"required":["id"]}}}},"RewardToUpdate":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reward"}}}},"TrashToAdd":{"content":{"application/json":{"schema":{"type":"object","properties":{"is_full":{"type":"boolean"},"nb_alerts":{"type":"integer"},"last_empty":{"type":"string","format":"date"},"qr_code":{"type":"string"},"position_id":{"type":"integer","$ref":"#/components/schemas/Position"}},"required":["is_full","nb_alerts","last_empty","qr_code","position_id"]}}}},"TrashToUpdate":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trash"}}}},"TrashToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"}},"required":["id"]}}}},"TrashToEmpty":{"content":{"application/json":{"schema":{"type":"object","properties":{"were_real_reports":{"type":"boolean"},"trash_id":{"type":"integer"}},"required":["were_real_reports","trash_id"]}}}},"Login":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["email","password"]}}}},"VendorToAdd":{"content":{"application/json":{"schema":{"type":"object","properties":{"name_fr":{"type":"string"},"name_en":{"type":"string"},"description_fr":{"type":"string"},"description_en":{"type":"string"},"throins_cost":{"type":"integer"},"real_cost":{"type":"number","format":"float"},"position_id":{"type":"integer","$ref":"#/components/schemas/Position"}},"required":["name_fr","name_en","description_fr","description_en","throins_cost","real_cost","position_id"]}}}},"VendorToDelete":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"}},"required":["id"]}}}},"VendorToUpdate":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vendor"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}